Let peerConn handle IPs in for tests
