+++ p2p/peer_test.go
@@ -1,6 +1,8 @@
 package p2p
 
 import (
+	golog "log"
+	"net"
 	"testing"
 	"time"
 
@@ -9,8 +11,11 @@ import (
 
 	crypto "github.com/tendermint/go-crypto"
 	tmconn "github.com/tendermint/tendermint/p2p/conn"
+	cmn "github.com/tendermint/tmlibs/common"
 	"github.com/tendermint/tmlibs/log"
 )
+
+const testCh = 0x01
 
 func TestPeerBasic(t *testing.T) {
 	assert, require := assert.New(t), require.New(t)
@@ -104,4 +109,78 @@ func createOutboundPeerAndPerformHandshake(addr *NetAddress, config *PeerConfig)
 	p := newPeer(pc, nodeInfo, reactorsByCh, chDescs, func(p Peer, r interface{}) {})
 	p.SetLogger(log.TestingLogger().With("peer", addr))
 	return p, nil
+}
+
+type remotePeer struct {
+	PrivKey  crypto.PrivKey
+	Config   *PeerConfig
+	addr     *NetAddress
+	quit     chan struct{}
+	channels cmn.HexBytes
+}
+
+func (rp *remotePeer) Addr() *NetAddress {
+	return rp.addr
+}
+
+func (rp *remotePeer) ID() ID {
+	return PubKeyToID(rp.PrivKey.PubKey())
+}
+
+func (rp *remotePeer) Start() {
+	l, e := net.Listen("tcp", "127.0.0.1:0") // any available address
+	if e != nil {
+		golog.Fatalf("net.Listen tcp :0: %+v", e)
+	}
+	rp.addr = NewNetAddress(PubKeyToID(rp.PrivKey.PubKey()), l.Addr())
+	rp.quit = make(chan struct{})
+	if rp.channels == nil {
+		rp.channels = []byte{testCh}
+	}
+	go rp.accept(l)
+}
+
+func (rp *remotePeer) Stop() {
+	close(rp.quit)
+}
+
+func (rp *remotePeer) accept(l net.Listener) {
+	conns := []net.Conn{}
+
+	for {
+		conn, err := l.Accept()
+		if err != nil {
+			golog.Fatalf("Failed to accept conn: %+v", err)
+		}
+
+		pc, err := newInboundPeerConn(conn, rp.Config, rp.PrivKey)
+		if err != nil {
+			golog.Fatalf("Failed to create a peer: %+v", err)
+		}
+
+		_, err = pc.HandshakeTimeout(NodeInfo{
+			ID:         rp.Addr().ID,
+			Moniker:    "remote_peer",
+			Network:    "testing",
+			Version:    "123.123.123",
+			ListenAddr: l.Addr().String(),
+			Channels:   rp.channels,
+		}, 1*time.Second)
+		if err != nil {
+			golog.Fatalf("Failed to perform handshake: %+v", err)
+		}
+
+		conns = append(conns, conn)
+
+		select {
+		case <-rp.quit:
+			for _, conn := range conns {
+				if err := conn.Close(); err != nil {
+					golog.Fatal(err)
+				}
+			}
+			return
+		default:
+		}
+	}
 }
